[KiRA]
Add new database: KiRA_Database.db
(under ApplicationFiles\DataBase)

New folders and files:

BusinessLogicLayer: Here you can find all programs, functions which are responsible for the proper communication between the GUI and the DataAccessLayer(DAL)

DataAccessLayer: It contains the programs, functions which are responsible for the direct communication with the SQLite database

GUI: It contains all windows forms

ValueObjects: It contains the models (eg. Person)

The .cs files will be described in the Wiki Page.